# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  donor: 'donor',
  organization: 'organization',
  admin: 'admin',
} as const;

export type RegistrationRole = typeof RegistrationRole[keyof typeof RegistrationRole];


export const RegistrationRole = {
  donor: 'donor',
  organization: 'organization',
} as const;

export type DonationCategory = typeof DonationCategory[keyof typeof DonationCategory];


export const DonationCategory = {
  electronics: 'electronics',
  plastics: 'plastics',
  glass: 'glass',
  paper: 'paper',
  metal: 'metal',
  clothing: 'clothing',
  food: 'food',
  other: 'other',
} as const;

export type DonationUnit = typeof DonationUnit[keyof typeof DonationUnit];


export const DonationUnit = {
  kg: 'kg',
  pieces: 'pieces',
  liters: 'liters',
} as const;

export type DonationStatus = typeof DonationStatus[keyof typeof DonationStatus];


export const DonationStatus = {
  pending: 'pending',
  approved: 'approved',
  delivered: 'delivered',
  rejected: 'rejected',
} as const;

export interface Error {
  error: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  organizationName: string | null;
  createdAt: string;
}

export interface UserSession {
  token: string;
  user: User;
}

export interface RegistrationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
  role: RegistrationRole;
  /**
     * @minLength 2
     * @maxLength 140
     */
  organizationName?: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  municipality?: string;
  /**
     * @minLength 10
     * @maxLength 700
     */
  description?: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  acceptedCategories?: DonationCategory[];
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Organization {
  id: string;
  name: string;
  municipality: string;
  description: string;
  acceptedCategories: DonationCategory[];
  isVerified: boolean;
}

export interface OrganizationVerificationInput {
  isVerified: boolean;
}

export interface Donation {
  id: string;
  donorId: string;
  donorName: string;
  recipientId: string;
  recipientName: string;
  category: DonationCategory;
  title: string;
  description: string;
  /** @minimum 0.01 */
  quantity: number;
  unit: DonationUnit;
  status: DonationStatus;
  createdAt: string;
  updatedAt: string;
}

export interface DonationInput {
  recipientId: string;
  category: DonationCategory;
  /**
     * @minLength 2
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 5
     * @maxLength 1000
     */
  description: string;
  /**
     * @maximum 1000000
     * @exclusiveMinimum 0
     */
  quantity: number;
  unit: DonationUnit;
}

export interface DonationStatusInput {
  status: DonationStatus;
}

export interface DashboardSummary {
  donationCount: number;
  pendingCount: number;
  approvedCount: number;
  deliveredCount: number;
  rejectedCount: number;
  /** Sum of registered donations measured in kilograms */
  totalWeightKg: number;
  recentDonations: Donation[];
}

export type ListDonationsParams = {
status?: DonationStatus;
};