# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Donation } from './donation';

export interface DashboardSummary {
  donationCount: number;
  pendingCount: number;
  approvedCount: number;
  deliveredCount: number;
  rejectedCount: number;
  /** Sum of registered donations measured in kilograms */
  totalWeightKg: number;
  recentDonations: Donation[];
}