# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DonationCategory } from './donationCategory';
import type { DonationStatus } from './donationStatus';
import type { DonationUnit } from './donationUnit';

export interface Donation {
  id: string;
  donorId: string;
  donorName: string;
  recipientId: string;
  recipientName: string;
  category: DonationCategory;
  title: string;
  description: string;
  /** @minimum 0.01 */
  quantity: number;
  unit: DonationUnit;
  status: DonationStatus;
  createdAt: Date;
  updatedAt: Date;
}