# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DonationCategory } from './donationCategory';
import type { DonationUnit } from './donationUnit';

export interface DonationInput {
  recipientId: string;
  category: DonationCategory;
  /**
     * @minLength 2
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 5
     * @maxLength 1000
     */
  description: string;
  /**
     * @maximum 1000000
     * @exclusiveMinimum 0
     */
  quantity: number;
  unit: DonationUnit;
}