# lib/api-zod/src/generated/types/donationStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DonationStatus } from './donationStatus';

export interface DonationStatusInput {
  status: DonationStatus;
}