# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DonationCategory } from './donationCategory';
import type { RegistrationRole } from './registrationRole';

export interface RegistrationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
  role: RegistrationRole;
  /**
     * @minLength 2
     * @maxLength 140
     */
  organizationName?: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  municipality?: string;
  /**
     * @minLength 10
     * @maxLength 700
     */
  description?: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  acceptedCategories?: DonationCategory[];
}