# lib/api-zod/src/generated/types/registrationRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RegistrationRole = typeof RegistrationRole[keyof typeof RegistrationRole];


export const RegistrationRole = {
  donor: 'donor',
  organization: 'organization',
} as const;